+++ frontend/portals/CartItemPrices/CartItemPrice/index.jsx
@@ -22,11 +22,11 @@ const arePricesEqual = (priceOne = {}, priceTwo = {}) => (
  */
 const CartItemPrice = ({
   subscriptionPrice,
-  subscriptions,
+  rechargeInfo,
   originalPrice,
   children,
 }) => {
-  if (!subscriptions.length || arePricesEqual(subscriptionPrice, originalPrice)) {
+  if (!rechargeInfo.length || arePricesEqual(subscriptionPrice, originalPrice)) {
     return children;
   }
   const { special: specialPrice, default: defaultPrice } = subscriptionPrice;
@@ -42,14 +42,14 @@ const CartItemPrice = ({
 CartItemPrice.propTypes = {
   children: PropTypes.node.isRequired,
   originalPrice: PropTypes.shape(),
+  rechargeInfo: PropTypes.arrayOf(PropTypes.shape()),
   subscriptionPrice: PropTypes.shape(),
-  subscriptions: PropTypes.arrayOf(PropTypes.shape()),
 };
 
 CartItemPrice.defaultProps = {
   originalPrice: {},
   subscriptionPrice: {},
-  subscriptions: [],
+  rechargeInfo: [],
 };
 
 export default connect(CartItemPrice);