+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.2.0] - 2018-11-18
+### Added
+- Display of recharge cart prices from Recharge cart response.
+### Fixed
+- Quantity issue with optional subscription products when adding non subscription and subscription products.
+
 ## [1.1.0] - 2018-11-08
 ### Added
 - Support for increasing quantity on non subscription products on the cart page